- moved TStoreWndDesc to core - TStore now supports GetWndDesc()
