+++ src/qminer/qminer_core.cpp
@@ -9,7 +9,7 @@
 #include "qminer_core.h"
 #include "qminer_ftr.h"
 #include "qminer_aggr.h"
-#include "geospatial_aggr.h"
+//#include "geospatial_aggr.h"
 
 namespace TQm {
 
@@ -135,6 +135,24 @@ void TNmValidator::AssertValidNm(const TStr& NmStr) const {
 
 void TNmValidator::SetStrictNmP(const bool& _StrictNmP) {
     StrictNmP = _StrictNmP;
+}
+
+///////////////////////////////
+// Store window description
+TStr TStoreWndDesc::SysInsertedAtFieldName = "_sys_inserted_at";
+
+void TStoreWndDesc::Save(TSOut& SOut) const {
+    TInt(WindowType).Save(SOut);
+    WindowSize.Save(SOut);
+    InsertP.Save(SOut);
+    TimeFieldNm.Save(SOut);
+}
+
+void TStoreWndDesc::Load(TSIn& SIn) {
+    WindowType = TStoreWndType(TInt(SIn).Val);
+    WindowSize.Load(SIn);
+    InsertP.Load(SIn);
+    TimeFieldNm.Load(SIn);
 }
 
 ///////////////////////////////
@@ -6525,7 +6543,7 @@ void TStreamAggr::Init() {
     Register<TStreamAggrs::TWinBufSpVecSum>();
     Register<TStreamAggrs::THistogramAD>();
     // geospatial aggregates
-    Register<TStreamAggrs::TStayPointDetector>();
+    //Register<TStreamAggrs::TStayPointDetector>();
 }
 
 TStreamAggr::TStreamAggr(const TWPt<TBase>& _Base, const TStr& _AggrNm): Base(_Base), AggrNm(_AggrNm) {
